upade the board model and user for relationship and make the user access to his Boards only

+++ app/Http/Controllers/BoardController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 use App\Models\Board;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BoardController extends Controller
 {
@@ -17,26 +18,31 @@ class BoardController extends Controller
     }
     public function index()
     {
-       return Board::all();
+       return Auth::user()->boards;
     }
 
     public function show($boardId)
         {
             $board = Board::findOrFail($boardId);
-           return $board;
+            if(Auth::user()->id !== $board->user_id){
+                return response()->json(['status' => 'error','message'=>'Unauthorized'], 401);
+            }
+                 return $board;
         }
 
     public function store(Request $request)
     {
-        Board::create([
+        Auth::user()->boards()->create([
             'name' => $request->name,
-            'user_id' => 1,
         ]);
         return response()->json(['message' => 'success'], 200);
     }
     public function update(Request $request, $boardId)
     {
         $board = Board::find($boardId);
+        if(Auth::user()->id !== $board->user_id){
+            return response()->json(['status' => 'error','message'=>'Unauthorized'], 401);
+        }
         $board->update($request->all());
         //$board->save();
 
@@ -44,6 +50,10 @@ class BoardController extends Controller
     }
     public function destroy($boardId)
     {
+        $board = Board::find($boardId);
+        if(Auth::user()->id !== $board->user_id){
+            return response()->json(['status' => 'error','message'=>'Unauthorized'], 401);
+        }
         if(Board::destroy($boardId)) {
             return response()->json(['status' => 'success','message'=>'Board Deleted successfully'], 200);
         }
